refactor:#265 client meal detail 리팩토링
