+++ src/app/(client)/meal/[id]/_components/meal-detail-section.tsx
@@ -7,21 +7,21 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { calculateTotalNutrition } from '@/utils/nutrition-calculator.util';
 import { MealCategory, MealCategoryType } from '@/types/meal-category.type';
-import TimePicker, { TimeFields } from '../../components/time-picker';
+import TimePicker, { TimeFields } from '../../_components/time-picker';
 import { updateMeal } from '@/apis/meal.api';
 import { formatTimestamp } from '@/utils/format.util';
 import { ERROR_MESSAGES } from '../../post/edit/constants/error-message.constant';
-import MealImageCarousel from '../../components/meal-images-carousel';
+import MealImageCarousel from '../../_components/meal-images-carousel';
 import TotalNutritionChart from './total-nutrition-chart';
 import MealList from './meal-list';
 import { Typography } from '@/components/ui/typography';
 import GlassBackground from '@/components/commons/glass-background';
-import { MEAL_CATEGORY } from '../../constants/category.constant';
+import { MEAL_CATEGORY } from '../../_constants/category.constant';
 import TagSelectItem from '@/components/commons/tag-select-item';
 import Responsive from '@/components/commons/responsive';
-import MealCalendarPc from '../../components/meal-calendar-pc';
-import MealCalendarDrawer from '../../components/meal-calendar-drawer';
-import TimePickerPc from '../../components/time-picker-pc';
+import MealCalendarPc from '../../_components/meal-calendar-pc';
+import MealCalendarDrawer from '../../_components/meal-calendar-drawer';
+import TimePickerPc from '../../_components/time-picker-pc';
 import MemoBox from './memo-box';
 import { Button } from '@/components/ui/button';
 import Modal from '@/components/commons/modal';
