+++ src/app/(client)/meal/[id]/layout.tsx
