+++ src/app/(client)/meal/_components/meal-images-carousel.tsx
