+++ src/app/(client)/meal/_components/time-picker.tsx
@@ -14,7 +14,7 @@ import { useState } from 'react';
 import Picker from 'react-mobile-picker';
 import { z } from 'zod';
 import { Button } from '@/components/ui/button';
-import { dateSchema } from '../detail/components/meal-detail-section';
+import { dateSchema } from '../[id]/_components/meal-detail-section';
 import { dateSelections } from '../post/edit/utils/time-picker.util';
 
 type TimePickerProps = {
